+++ paper/src/main/java/com/denizenscript/denizen/paper/PaperModule.java
@@ -75,6 +75,7 @@ public class PaperModule {
             ScriptEvent.registerScriptEvent(PlayerRaiseLowerItemScriptEventPaperImpl.class);
         }
         ScriptEvent.registerScriptEvent(PlayerSelectsStonecutterRecipeScriptEvent.class);
+        ScriptEvent.registerScriptEvent(PlayerSetSpawnScriptEvent.class);
         if (NMSHandler.getVersion().isAtLeast(NMSVersion.v1_20)) {
             ScriptEvent.registerScriptEvent(PlayerShieldDisableScriptEvent.class);
         }
