+++ plugin/src/main/java/com/denizenscript/denizen/events/ScriptEventRegistry.java
@@ -247,9 +247,6 @@ public class ScriptEventRegistry {
         ScriptEvent.registerScriptEvent(PlayerShearsScriptEvent.class);
         ScriptEvent.registerScriptEvent(PlayerSmithsItemScriptEvent.class);
         ScriptEvent.registerScriptEvent(PlayerSneakScriptEvent.class);
-        if (NMSHandler.getVersion().isAtLeast(NMSVersion.v1_19)) {
-            ScriptEvent.registerScriptEvent(PlayerSpawnChangeScriptEvent.class);
-        }
         ScriptEvent.registerScriptEvent(PlayerSprintScriptEvent.class);
         ScriptEvent.registerScriptEvent(PlayerStandsOnScriptEvent.class);
         ScriptEvent.registerScriptEvent(PlayerStatisticIncrementsScriptEvent.class);
